Fix static analysis issues
:Release Notes:
Fix static analysis issues

:Detailed Notes:
Fix static analysis issues

:Testing Performed:
Local Test

:QA Notes:
N/A

:Issues Addressed:
[WRP-22928] Fix static analysis issues

Reviewed-on: http://gpro.lge.com/c/webosose/edge-ai-computer-vision/+/362842
Tested-by: Seungchul Han <seungchul.han@lge.com>

+++ framework/base/DetectorParam.cpp
@@ -63,7 +63,7 @@ t_aif_status DetectorParam::fromJson(const std::string &param) {
 
 bool DetectorParam::getUseXnnpack() const { return m_useXnnpack; }
 
-size_t DetectorParam::getNumThreads() const { return m_numThreads; }
+int DetectorParam::getNumThreads() const { return m_numThreads; }
 
 const std::vector<std::pair<std::string, std::string>> &
 DetectorParam::getDelegates() const {
